+++ mcmd/commands/add.py
@@ -4,8 +4,8 @@ from pathlib import Path
 
 import mcmd.config.config as config
 from mcmd import io
-from mcmd.client import api
-from mcmd.client.molgenis_client import post, get, post_files
+from mcmd.molgenis import api
+from mcmd.molgenis.client import post, get, post_files
 from mcmd.command import command
 from mcmd.commands._registry import arguments
 from mcmd.io import highlight
