+++ mcmd/commands/delete.py
@@ -1,12 +1,12 @@
 from urllib.parse import urljoin
 
-import mcmd.client.molgenis_client as client
+import mcmd.molgenis.client as client
 from mcmd import io
-from mcmd.client import api
+from mcmd.molgenis import api
 from mcmd.command import command
 from mcmd.commands._registry import arguments
 from mcmd.io import highlight
-from mcmd.utils.resources import detect_resource_type, ensure_resource_exists, ResourceType
+from mcmd.molgenis.resources import detect_resource_type, ensure_resource_exists, ResourceType
 
 
 # =========
