+++ mcmd/commands/enable.py
@@ -1,11 +1,11 @@
 from mcmd import io
-from mcmd.client import api
-from mcmd.client.molgenis_client import post
+from mcmd.molgenis import api
+from mcmd.molgenis.client import post
 from mcmd.command import command
 from mcmd.commands._registry import arguments
 from mcmd.io import highlight
 from mcmd.utils.errors import McmdError
-from mcmd.utils.resources import one_resource_exists, ensure_resource_exists, ResourceType
+from mcmd.molgenis.resources import one_resource_exists, ensure_resource_exists, ResourceType
 
 
 # =========
