+++ mcmd/commands/ping.py
@@ -1,11 +1,11 @@
 from colorama import Fore
 
 import mcmd.config.config as config
-from mcmd.command import command
+import mcmd.molgenis.version as molgenis_version
 from mcmd.commands._registry import arguments
-from mcmd.io import highlight
-from mcmd.utils.errors import McmdError
-from mcmd.version import molgenis_version
+from mcmd.core.command import command
+from mcmd.core.errors import McmdError
+from mcmd.io.io import highlight
 
 
 # =========
