+++ mcmd/commands/set.py
@@ -4,8 +4,8 @@ Alter settings
 import json
 
 from mcmd import io
-from mcmd.client import api
-from mcmd.client.molgenis_client import get, put
+from mcmd.molgenis import api
+from mcmd.molgenis.client import get, put
 from mcmd.command import command
 from mcmd.commands._registry import arguments
 from mcmd.io import highlight
