+++ mcmd/molgenis/auth.py
@@ -10,7 +10,7 @@ import requests
 from requests import HTTPError
 
 from mcmd import io
-from mcmd.client import api
+from mcmd.molgenis import api
 from mcmd.config import config
 from mcmd.utils.errors import McmdError, MolgenisOfflineError
