+++ mcmd/molgenis/resources.py
@@ -1,8 +1,8 @@
 from enum import Enum
 from typing import List
 
-from mcmd.client import api
-from mcmd.client.molgenis_client import get
+from mcmd.molgenis import api
+from mcmd.molgenis.client import get
 from mcmd.io import multi_choice
 from mcmd.logging import get_logger
 from mcmd.utils.errors import McmdError
